requirements.txt
fastapi>=0.128.0
huggingface-hub>=1.3.1
numpy>=2.4.1
onnxruntime>=1.23.2
opencv-python>=4.11.0.86
pdfplumber>=0.11.9
pillow>=12.1.0
pydantic-settings>=2.12.0
pymupdf>=1.26.7
python-multipart>=0.0.21
rapidocr-onnxruntime>=1.4.0
uvicorn>=0.40.0
